Parameterized test suite.
