+++ openmdao/test_suite/test_general.py
@@ -1,6159 +1,105 @@
-
 from __future__ import division, print_function
 
+import itertools
 import unittest
+from six.moves import range
+import numpy
 
-from openmdao.test_suite.update_test_general import CompTestCaseBase
-from openmdao.test_suite.components.implicit_components     import TestImplCompNondLinear
-from openmdao.test_suite.components.explicit_components     import TestExplCompNondLinear
+from openmdao.test_suite.components.implicit_components \
+    import TestImplCompNondLinear
+from openmdao.test_suite.components.explicit_components \
+    import TestExplCompNondLinear
+from openmdao.test_suite.groups.group import TestGroupFlat
+from openmdao.api import Problem
 from openmdao.api import DefaultVector, NewtonSolver, ScipyIterativeSolver
+from openmdao.api import GlobalJacobian, DenseMatrix, CooMatrix, CsrMatrix
 from openmdao.parallel_api import PETScVector
 
-class CompTestCase(CompTestCaseBase):
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_dense_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_dense_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c1_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c1_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c1_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c1_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c2_s1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (1,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c2_s2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (2,))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c2_s2x1(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestImplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c2_s1x2(self):
-        self.run_test(TestImplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_matvec_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'matvec', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_dense_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'dense', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_coo_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_implicit_sparse_csr_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_matvec_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'matvec', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_dense_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'dense', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_coo_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_DefaultVector_explicit_sparse_csr_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, DefaultVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_matvec_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'matvec', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_dense_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'dense', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_coo_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-coo', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_implicit_sparse_csr_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'implicit', 'sparse-csr', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_matvec_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'matvec', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_dense_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'dense', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_coo_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-coo', 'aij', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_array_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'array', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_sparse_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'sparse', 2, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v1_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 1, 2, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c1_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c1_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c1_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c1_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 1, (1, 2))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c2_s1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (1,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c2_s2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (2,))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c2_s2x1(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (2, 1))
-
-
-    def test_TestExplCompNondLinear_PETScVector_explicit_sparse_csr_aij_v2_c2_s1x2(self):
-        self.run_test(TestExplCompNondLinear, PETScVector, 'explicit', 'sparse-csr', 'aij', 2, 2, (1, 2))
-
-
-if __name__ == '__main__':
-    unittest.main()
+from nose_parameterized import parameterized
+
+def custom_name(testcase_func, param_num, param):
+    return 'test' + '_'.join(p.__name__ for p in param.args[:2]) + '_'.join(str(p) for p in param.args[2:])
+
+class CompTestCaseBase(unittest.TestCase):
+    """The TestCase that actually runs all of the cases inherits from this."""
+
+    @parameterized.expand(itertools.product(
+        [TestImplCompNondLinear, TestExplCompNondLinear],
+        [DefaultVector, PETScVector],
+        ['implicit', 'explicit'],
+        ['matvec', 'dense', 'sparse-coo', 'sparse-csr'],
+        ['array', 'sparse', 'aij'],
+        range(1, 3),
+        range(1, 3),
+        [(1,), (2,), (2, 1), (1, 2)],
+    ), testcase_func_name=custom_name)
+    def test_openmdao(self, Component, Vector, connection_type, jacobian_type,
+                 partial_type, num_var, num_comp, var_shape):
+
+        group = TestGroupFlat(num_comp=num_comp, num_var=num_var,
+                              var_shape=var_shape,
+                              connection_type=connection_type,
+                              jacobian_type=jacobian_type,
+                              partial_type=partial_type,
+                              Component=Component,
+                              )
+        prob = Problem(group).setup(Vector)
+
+        if jacobian_type == 'dense':
+            prob.root.jacobian = GlobalJacobian(Matrix=DenseMatrix)
+        elif jacobian_type == 'sparse-coo':
+            prob.root.jacobian = GlobalJacobian(Matrix=CooMatrix)
+        elif jacobian_type == 'sparse-csr':
+            prob.root.jacobian = GlobalJacobian(Matrix=CsrMatrix)
+
+        prob.root.nl_solver = NewtonSolver(
+            subsolvers={'linear': ScipyIterativeSolver(
+                maxiter=100,
+            )}
+        )
+        prob.root.ln_solver = ScipyIterativeSolver(
+            maxiter=200, atol=1e-10, rtol=1e-10)
+        prob.root.suppress_solver_output = True
+
+        fail, rele, abse = prob.run()
+        if fail:
+            self.fail('Problem run failed: re %f ; ae %f' % (rele, abse))
+
+        # Setup for the 4 tests that follow
+        size = numpy.prod(var_shape)
+        work = prob.root._vectors['output']['']._clone()
+        work.set_const(1.0)
+        if Component == TestImplCompNondLinear:
+            val = 1 - 0.01 + 0.01 * size * num_var * num_comp
+        if Component == TestExplCompNondLinear:
+            val = 1 - 0.01 * size * num_var * (num_comp - 1)
+
+        # 1. fwd apply_linear test
+        prob.root._vectors['output'][''].set_const(1.0)
+        prob.root._apply_linear([''], 'fwd')
+        prob.root._vectors['residual'][''].add_scal_vec(-val, work)
+        self.assertAlmostEqual(
+            prob.root._vectors['residual'][''].get_norm(), 0)
+
+        # 2. rev apply_linear test
+        prob.root._vectors['residual'][''].set_const(1.0)
+        prob.root._apply_linear([''], 'rev')
+        prob.root._vectors['output'][''].add_scal_vec(-val, work)
+        self.assertAlmostEqual(
+            prob.root._vectors['output'][''].get_norm(), 0)
+
+        # 3. fwd solve_linear test
+        prob.root._vectors['output'][''].set_const(0.0)
+        prob.root._vectors['residual'][''].set_const(val)
+        prob.root._solve_linear([''], 'fwd')
+        prob.root._vectors['output'][''] -= work
+        self.assertAlmostEqual(
+            prob.root._vectors['output'][''].get_norm(), 0, delta=1e-2)
+
+        # 4. rev solve_linear test
+        prob.root._vectors['residual'][''].set_const(0.0)
+        prob.root._vectors['output'][''].set_const(val)
+        prob.root._solve_linear([''], 'rev')
+        prob.root._vectors['residual'][''] -= work
+        self.assertAlmostEqual(
+            prob.root._vectors['residual'][''].get_norm(), 0, delta=1e-2)
