+++ setup.py
@@ -46,6 +46,7 @@ setup(name='openmdao',
         'six', 'numpydoc', #'numpy>=1.9.2',
         'scipy', 'sqlitedict',
         'pep8', 'pep257',
+        'nose-parameterized',
       ],
       #scripts=['bin/om-pylint.sh']
       entry_points="""
